atualização

+++ Exercício 1 - Cálculo de área de um retângulo.py
@@ -6,7 +6,7 @@ comprimento_retangulo = input('Insira o comprimento do retângulo:')
 largura_retangulo = input('Insira a largura do retângulo:')
 
 # Cálculo da área #
-area = int(comprimento_retangulo) * int(largura_retangulo)
+area = float(comprimento_retangulo) * float(largura_retangulo)
 
 # Imprime o resultado na tela do usuário
 print(area)
